+++ lib/ui/screens/overview/components/printer_card_controller.dart
@@ -3,6 +3,9 @@
  * All rights reserved.
  */
 
+import 'dart:math';
+
+import 'package:collection/collection.dart';
 import 'package:common/data/dto/server/klipper.dart';
 import 'package:common/data/model/hive/machine.dart';
 import 'package:common/data/model/moonraker_db/webcam_info.dart';
@@ -11,6 +14,7 @@ import 'package:common/service/moonraker/klippy_service.dart';
 import 'package:common/service/moonraker/webcam_service.dart';
 import 'package:common/service/payment_service.dart';
 import 'package:common/service/selected_machine_service.dart';
+import 'package:common/service/setting_service.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mobileraker/routing/app_router.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
@@ -42,14 +46,29 @@ class PrinterCardController extends _$PrinterCardController {
 
     if (klipperState != KlipperState.ready) return;
 
-    var filteredCamsFuture = ref.watch(allSupportedWebcamInfosProvider(machine.uuid).future);
-    if (!ref.watch(isSupporterProvider)) {
-      filteredCamsFuture = filteredCamsFuture
-          .then((value) => value.where((element) => !element.service.forSupporters).toList(growable: false));
+    var cams = await ref.watch(allSupportedWebcamInfosProvider(machine.uuid).future);
+    if (cams.isEmpty) {
+      yield null;
+      return;
     }
 
-    var filteredCams = await filteredCamsFuture;
-    yield filteredCams.firstOrNull;
+    var webcamIndexKey = CompositeKey.keyWithString(UtilityKeys.webcamIndex, machine.uuid);
+    var selIndex = min(
+      cams.length - 1,
+      max(0, ref.read(settingServiceProvider).readInt(webcamIndexKey, 0)),
+    );
+
+    WebcamInfo? previewCam = cams.elementAtOrNull(selIndex);
+    var isSupporter = ref.watch(isSupporterProvider);
+
+    // If there is no preview cam or the preview cam is not for supporters or the user is a supporter
+    if (previewCam == null || previewCam.service.forSupporters == false || isSupporter) {
+      yield previewCam;
+      return;
+    }
+
+    // If the user is not a supporter and the cam he selected is for supporters only just select the first cam that is not for supporters
+    yield cams.firstWhereOrNull((element) => element.service.forSupporters == false);
   }
 
   onTapTile() {